requirements.txt
Flask==1.0.3
pylint==2.3.1
python-telegram-bot[socks]
gunicorn